online-shop-backend/online_shop/requirements.txt
asgiref==3.8.1
DateTime==5.5
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.3
tzdata==2024.2
uuid==1.30
zope.interface==7.2
